+++ freebsd5/sys/dev/ppbus/vpo.c
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $FreeBSD: src/sys/dev/ppbus/vpo.c,v 1.24 2002/04/07 22:06:20 ticso Exp $
+ * $FreeBSD: src/sys/dev/ppbus/vpo.c,v 1.28 2003/05/01 19:46:21 jhb Exp $
  */
 
 #include <sys/param.h>
@@ -31,8 +31,6 @@
 #include <sys/module.h>
 #include <sys/bus.h>
 #include <sys/malloc.h>
-#include <sys/devicestat.h>	/* for struct devstat */
-
 
 #include <cam/cam.h>
 #include <cam/cam_ccb.h>
@@ -359,10 +357,10 @@ vpo_action(struct cam_sim *sim, union ccb *ccb)
 		ccg = &ccb->ccg;
 
 #ifdef VP0_DEBUG
-		printf("vpo%d: XPT_CALC_GEOMETRY (bs=%d,vs=%d,c=%d,h=%d,spt=%d) request\n",
+		printf("vpo%d: XPT_CALC_GEOMETRY (bs=%d,vs=%jd,c=%d,h=%d,spt=%d) request\n",
 			vpo->vpo_unit,
 			ccg->block_size,
-			ccg->volume_size,
+			(intmax_t)ccg->volume_size,
 			ccg->cylinders,
 			ccg->heads,
 			ccg->secs_per_track);
